+++ app/services/draw_size_restrict_toggler.rb
@@ -4,13 +4,13 @@
 # Service object to toggle whether or not a group size is locked for a given
 # draw. Validates that the size is one for which there are currently available
 # suites.
-class DrawSizeLockToggler
+class DrawSizeRestrictToggler
   include ActiveModel::Model
   include Callable
 
   validate :size_is_an_integer_string
 
-  # Initialize a new DrawSizeLockToggler
+  # Initialize a new DrawSizeRestrictToggler
   #
   # @param draw [Draw] the draw in question
   # @param size [String] the size param passed in the request
@@ -21,7 +21,7 @@ class DrawSizeLockToggler
 
   def toggle
     return error(self) unless valid?
-    toggle_size_lock
+    toggle_size_restrict
     draw.save!
     success
   rescue ActiveRecord::RecordInvalid => e
@@ -42,11 +42,11 @@ class DrawSizeLockToggler
     end
   end
 
-  def toggle_size_lock
-    if draw.size_locked?(size)
-      draw.locked_sizes.delete_if { |s| s == size }
+  def toggle_size_restrict
+    if draw.size_restricted?(size)
+      draw.restricted_sizes.delete_if { |s| s == size }
     else
-      draw.locked_sizes << size
+      draw.restricted_sizes << size
     end
   end
 
@@ -61,10 +61,10 @@ class DrawSizeLockToggler
 
   def success_msg
     size_str = Suite.size_str(size).pluralize.capitalize
-    if draw.reload.size_locked?(size)
-      "#{size_str} locked"
+    if draw.reload.size_restricted?(size)
+      "#{size_str} restricted."
     else
-      "#{size_str} unlocked"
+      "#{size_str} permitted."
     end
   end
 end