+++ db/migrate/20181114215106_change_locked_sizes_to_restricted_sizes.rb
@@ -1,0 +1,5 @@
+class ChangeLockedSizesToRestrictedSizes < ActiveRecord::Migration[5.1]
+  def change
+  	rename_column :draws, :locked_sizes, :restricted_sizes
+  end
+end