+++ spec/features/draws/draw_locking_sizes_spec.rb
@@ -2,14 +2,14 @@
 
 require 'rails_helper'
 
-RSpec.feature 'Draw suite size locking' do
+RSpec.feature 'Draw suite size restricting' do
   before { log_in create(:admin) }
   let(:draw) do
     create(:draw, status: 'group_formation').tap do |d|
       d.suites << Array.new(3) do |i|
         create(:suite_with_rooms, rooms_count: i + 1)
       end
-      d.locked_sizes << 1
+      d.restricted_sizes << 1
     end
   end
 
@@ -18,20 +18,20 @@ RSpec.feature 'Draw suite size locking' do
     check('Doubles')
     uncheck('Singles')
     click_on 'Save'
-    expect(draw.reload.locked_sizes).to eq([2])
+    expect(draw.reload.restricted_sizes).to eq([2])
   end
 
-  describe 'lock all sizes' do
+  describe 'restrict all sizes' do
     it 'can can be done from the draw page' do
       visit draw_path(draw)
-      click_on 'Lock all sizes'
+      click_on 'Restrict all sizes'
       expect(page).to have_css('.flash-notice', text: "#{draw.name} updated")
     end
 
     it 'works' do
       visit draw_path(draw)
-      click_on 'Lock all sizes'
-      expect(draw.reload.locked_sizes).to eq(draw.suite_sizes)
+      click_on 'Restrict all sizes'
+      expect(draw.reload.restricted_sizes).to eq(draw.suite_sizes)
     end
   end
 end