+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Changed
+* Rename group size locking to restrict size locking ([811](https://gitlab.com/yale-sdmp/vesta/issues/811)).
+
 ## v2.2.1 - 2019-02-20
 ### Fixed
 * Fix the display of archived group links on the student dashboard ([#907](https://gitlab.com/yale-sdmp/vesta/issues/907)).